requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.33.13
botocore==1.33.13
certifi==2026.1.4
charset-normalizer==3.4.4
Django==3.2.25
django-cors-headers==4.1.0
django-ses==3.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
idna==3.10
importlib-metadata==6.7.0
jmespath==1.0.1
mysqlclient @ file:///C:/Users/HANIMA/mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl
openpyxl==3.1.3
Pillow==9.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.31.0
s3transfer==0.8.2
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.20
zipp==3.15.0
